// src/types/course.ts
export interface Course {
  id: string;
  title: string;
  description: string;
  duration: string;
  students: string;
  rating: number;
  image: string;
  price: number;
  instructor: string;
  fullDescription: string;
  learningOutcomes: string[];
  requirements: string[];
  includes: string[];
}